# public/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AIGens API Documentation",
    "version": "1.0.0",
    "description": "Documentazione delle API per la piattaforma AIGens",
    "contact": {
      "name": "Supporto AIGens",
      "email": "support@aigens.io",
      "url": "https://aigens.io/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5555",
      "description": "Server di sviluppo locale"
    },
    {
      "url": "https://api.aigens.io",
      "description": "Server di produzione"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login con email e password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email dell'utente"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password dell'utente"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login effettuato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object"
                        },
                        "token": {
                          "type": "string",
                          "description": "Token JWT per l'autenticazione API"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore nella validazione dell'email o password"
          },
          "401": {
            "description": "Credenziali non valide"
          },
          "404": {
            "description": "Utente non trovato"
          },
          "500": {
            "description": "Errore del server"
          }
        }
      }
    },
    "/api/v1/auth/login/passwordless": {
      "post": {
        "summary": "Richiede un link di accesso via email (autenticazione passwordless)",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email dell'utente"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Link di accesso inviato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Link di accesso inviato all'email fornita"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore nella validazione dell'email"
          },
          "429": {
            "description": "Troppe richieste, riprova più tardi"
          },
          "500": {
            "description": "Errore del server"
          }
        }
      }
    },
...
    "/api/v1/auth/refresh-token": {
      "post": {
        "summary": "Rinnova il token JWT",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token rinnovato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "Nuovo token JWT"
                        },
                        "expiresIn": {
                          "type": "string",
                          "description": "Durata del token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autorizzato"
          },
          "500": {
            "description": "Errore del server"
          }
        }
      }
    },
...
    "/api/v1/stripe/webhook": {
      "post": {
        "summary": "Webhook Stripe per gestire eventi di pagamento",
        "tags": [
          "Stripe"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento gestito con successo"
          },
          "400": {
            "description": "Errore nel webhook"
          }
        }
      }
    },
...
    "/api/v1/waitlist/subscribe": {
      "post": {
        "summary": "Aggiunge un'email alla lista d'attesa",
        "tags": [
          "Waitlist"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email da aggiungere alla lista d'attesa",
                    "example": "user@example.com"
                  },
                  "firstname": {
                    "type": "string",
                    "description": "Nome del contatto (opzionale)",
                    "example": "Mario"
                  },
                  "lastname": {
                    "type": "string",
                    "description": "Cognome del contatto (opzionale)",
                    "example": "Rossi"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email aggiunta con successo alla lista d'attesa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Email user@example.com aggiunta con successo alla lista d'attesa"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        },
                        "listName": {
                          "type": "string",
                          "example": "Aigens.io Waiting List"
                        },
                        "listId": {
                          "type": "string",
                          "example": 159023
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Richiesta non valida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Email già presente nella lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "L'email è già presente nella lista d'attesa"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}